Add two files
