+++ modules/variables.js
@@ -1,0 +1,11 @@
+const contactLink = document.querySelector('.contact-us');
+const listOfBooks = document.querySelector('.container');
+const booksection = document.querySelector('.add-book');
+const listLink = document.querySelector('.list');
+const contactSection = document.querySelector('.contact');
+const addNewLink = document.querySelector('.add-new');
+const titleBookOne = document.querySelector('.books-title');
+
+export {
+  contactLink, listOfBooks, booksection, listLink, contactSection, addNewLink, titleBookOne,
+};
